requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Time series forecasting (optional - uncomment if needed)
# prophet>=1.1.0
# statsmodels>=0.14.0

# CLI and utilities
rich>=13.0.0  # Beautiful terminal formatting
click>=8.1.0  # CLI framework

# Development
pytest>=7.4.0
black>=23.0.0
pylint>=2.17.0
ipython>=8.12.0
jupyter>=1.0.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0.0

# Phase 5: Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
